move DryRunClient, more config
